Fix database query for slug. Fix auth check before deleting a post. Fix nullable on slug column in posts table.

+++ app/Repositories/post/EloquentPostRepository.php
@@ -193,7 +193,7 @@ class EloquentPostRepository extends CRepository implements PostRepository {
 	// tar bort ett block.
 	public function delete($id){
 		$Post = Post::find($id);
-		if($Post != null){
+		if($Post != null && Auth::check()){
 			if($Post->user_id == Auth::user()->id){
 				return $Post->delete();
 			}
